requirements.txt
streamlit
streamlit-drawable-canvas
tensorflow
numpy
pillow
matplotlib